fix: Corregido un problema a la hora de importar eventos
No se podían importar eventos con horas con decimales. Ya está cambiado.

+++ database/migrations/develop/2020_07_16_185500_create_event_table.php
@@ -23,7 +23,7 @@ class CreateEventTable extends Migration
             $table->integer('capacity');
             $table->enum('status',['draft','live','started','ended','completed','canceled']);
             $table->string('url');
-            $table->integer('hours')->nullable();
+            $table->float('hours')->nullable();
             $table->timestamps();
         });
     }
